Split ReshapeNode and SizeNode into a new nodes/manipulation.hpp
This makes indexing.hpp/indexing.cpp a bit smaller.
