+++ dwave/optimization/include/dwave-optimization/nodes/indexing.hpp
@@ -26,6 +26,7 @@
 
 #include "dwave-optimization/array.hpp"
 #include "dwave-optimization/graph.hpp"
+#include "dwave-optimization/state.hpp"
 
 namespace dwave::optimization {
 
@@ -272,51 +273,6 @@ class PermutationNode : public ArrayOutputMixin<ArrayNode> {
     // pointers to the "array" part of the two predecessors
     const Array* array_ptr_;
     const Array* order_ptr_;
-};
-
-class ReshapeNode : public ArrayOutputMixin<ArrayNode> {
- public:
-    ReshapeNode(ArrayNode* node_ptr, std::span<const ssize_t> shape);
-    ReshapeNode(ArrayNode* array_ptr, std::vector<ssize_t>&& shape);
-
-    double const* buff(const State& state) const override;
-    void commit(State& state) const override;
-    std::span<const Update> diff(const State& state) const override;
-    void revert(State& state) const override;
-
- private:
-    // we could dynamically cast each time, but it's easier to just keep separate
-    // pointer to the "array" part of the predecessor
-    const Array* array_ptr_;
-};
-
-class SizeNode : public ScalarOutputMixin<ArrayNode> {
- public:
-    explicit SizeNode(ArrayNode* node_ptr);
-
-    double const* buff(const State& state) const override;
-
-    void commit(State& state) const override;
-
-    std::span<const Update> diff(const State&) const override;
-
-    void initialize_state(State& state) const override;
-
-    // SizeNode's value is always a non-negative integer.
-    bool integral() const override { return true; }
-
-    double max() const override;
-
-    double min() const override;
-
-    void propagate(State& state) const override;
-
-    void revert(State& state) const override;
-
- private:
-    // we could dynamically cast each time, but it's easier to just keep separate
-    // pointer to the "array" part of the predecessor
-    const Array* array_ptr_;
 };
 
 }  // namespace dwave::optimization
